Make utern print logs as many as possible
The responses from FilterLogEvents API include only logs with
a timestamp after the value `startTime` but utern uses the last
ingestion time of the last response, which is usually larger
than the last event time, that is, utern ignores some logs that
occurred before the last ingestion time.
This commit makes utern use the minimum value of event times to
print logs as much as possible.